Exclude LuxRender GUI-app from being killed when closing renderview in Blende or Blender itself ( workaround, sorry Doug )

+++ src/luxrender/core/__init__.py
@@ -484,7 +484,7 @@ class RENDERENGINE_luxrender(bpy.types.RenderEngine, engine_base):
 				
 				# If we exit the wait loop (user cancelled) and renderer still running, then can it
 				# TODO: this should be a user option
-				if luxrender_process.poll() == None:
+				if luxrender_process.poll() == None and binary_name != 'luxrender':
 					luxrender_process.terminate()
 				
 				from luxrender.export.film import resolution
